pyproject.toml
[project]
name = "crewai"
version = "0.70.2"
description = "Cutting-edge framework for orchestrating role-playing, autonomous AI agents. By fostering collaborative intelligence, CrewAI empowers agents to work together seamlessly, tackling complex tasks."
readme = "README.md"
requires-python = ">=3.10,<=3.13"
authors = [
    { name = "Joao Moura", email = "joao@crewai.com" }
]
dependencies = [
    "pydantic>=2.4.2",
    "langchain>=0.2.16",
    "openai>=1.13.3",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp-proto-http>=1.22.0",
    "instructor>=1.3.3",
    "regex>=2024.9.11",
    "crewai-tools>=0.12.1",
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
    "appdirs>=1.4.4",
    "jsonref>=1.1.0",
    "agentops>=0.3.0",
    "embedchain>=0.1.114",
    "mem0ai>=0.1.20a0",
    "json-repair>=0.25.2",
    "auth0-python>=4.7.1",
    "litellm>=1.44.22",
    "pyvis>=0.3.2",
    "uv>=0.4.18",
    "tomli-w>=1.1.0",
]

[project.urls]
Homepage = "https://crewai.com"
Documentation = "https://docs.crewai.com"
Repository = "https://github.com/crewAIInc/crewAI"

[project.optional-dependencies]
tools = ["crewai-tools>=0.12.1"]
agentops = ["agentops>=0.3.0"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.4.10",
    "mypy>=1.10.0",
    "pre-commit>=3.6.0",
    "mkdocs>=1.4.3",
    "mkdocstrings>=0.22.0",
    "mkdocstrings-python>=1.1.2",
    "mkdocs-material>=9.5.7",
    "mkdocs-material-extensions>=1.3.1",
    "pillow>=10.2.0",
    "cairosvg>=2.7.1",
    "crewai-tools>=0.12.1",
    "pytest>=8.0.0",
    "pytest-vcr>=1.0.2",
    "python-dotenv>=1.0.0",
    "pytest-asyncio>=0.23.7",
    "pytest-subprocess>=1.5.2",
]

[project.scripts]
crewai = "crewai.cli.cli:crewai"

[tool.mypy]
ignore_missing_imports = true
disable_error_code = 'import-untyped'
exclude = ["cli/templates"]

[tool.bandit]
exclude_dirs = ["src/crewai/cli/templates"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
